Home

BPJSpec Documentation

BPJSpec is a 21-cm power spectrum code developed for the OVRO-LWA based on the $m$-mode analysis formalism.

BPJSpec.TransferMatrix — Type.

struct TransferMatrix <: AbstractBlockMatrix

This type represents the transfer matrix of an interferometer. This matrix effectively describes how an interferometer responds to the sky, including the antenna primary beam, bandpass, and baseline distribution.

This matrix is hierarchical in the sense that we save on some computational and storage requirements by separating long baselines from short baselines.

Fields

metadata stores the interferometer's metadata

storage contains instructions on how to read the matrix from disk

lmax is the largest value of the $l$ quantum number used by the matrix

mmax is the largest value of the $m$ quantum number used by the matrix

source

BPJSpec.MModes — Type.

struct MModes{S} <: AbstractBlockMatrix

This type represents the m-modes measured by the interferometer.

Fields

matrix is the BlockMatrix backend that stores the m-modes

source

Internals

BPJSpec.BlockMatrix — Type.

struct BlockMatrix{B, M, S, N} <: AbstractBlockMatrix

This type represents a (potentially enormous) block-diagonal matrix. This type is designed to be general enough to handle large matrices that fit in memory as well as enormous matrices that do not fit in memory. In principle this type can also be used to store small matrices, but it would be relatively inefficient compared to the standard Array{T, N}.

Note

Generally speaking the user should not use this type directly, but rather one of the numerous wrappers around this type.

Type Parameters

B specifies the type of the blocks that compose the matrix

N specifies the number of indices used to index into the matrix blocks

M specifies the type of metadata that is associated with the matrix

S specifies the type of storage used for the matrix

Fields

metadata stores any extra information we want associated with the matrix (eg. frequency channels or antenna positions).

storage contains instructions on how to read matrix blocks

cache is used if we want to read the matrix from disk and then keep it in memory for faster access.

source