Imaging

Imaging

API

BPJSpec.tikhonov — Function.

tikhonov(transfermatrix, mmodes; regularization=1e-2, mfs=false)

Create a dirty image of the sky using Tikhonov regularization.

Arguments:

transfermatrix the interferometer's transfer matrix, describing its response to the sky

mmodes the $m$-modes measured by the interferometer

Keyword Arguments:

regularization the amplitude of the Tikhonov regularization parameter

mfs determines whether or not to perform Multi-Frequency Synthesis imaging. If this parameter is set to true, all frequency channels will be used to generate a single image of the sky. If this parameter is set to false, an image of the sky will be generated for each frequency channel.

source