Noise Model

Noise Model

API

BPJSpec.NoiseModel — Type.

struct NoiseModel

This type represents the thermal noise contributed to the measurement of a set of $m$-modes.

A careful reading of Taylor, Carilli, Perley chapter 9 reveals that under the convention that Stokes-I is $(\rm xx + yy)/2$ we get the following expressions:

Uncertainty on single polarization visibilities (in flux density units):

\[σ_{\rm xx} = \frac{\sqrt{2} k T_{\rm sys}}{A_e \sqrt{Δν\,τ}}\]

Uncertainty on Stokes I visibilities (in flux density units):

\[σ_\text{Stokes-I} = \frac{σ_{\rm xx}}{\sqrt{2}} = \frac{k T_{\rm sys}}{A_e \sqrt{Δν\,τ}}\]

Where $k$ is the Boltzmann constant, $T_{\rm sys}$ is the system temperature, $A_e$ is the effective collecting area, $Δν$ is the bandwidth, and $τ$ is the integration time.

However, for a dipole antenna, the effective collecting area is not a very physically meaningful value. However, it turns out that we can relate the effective collecting are to the solid angle subtended by the primary beam $Ω$:

\[A_e = \frac{λ^2}{Ω}\]

Note

There seems to be some ambiguity in the literature in regards to notation. I believe we originally assumed that $A_e$ refers to the maximum effective collecting area, and that we have normalized the beam to be unity in that direction.

Finally we end up with the following expression after including an additional contribution due to time smearing:

\[σ_\text{m-modes} = \frac{k T_{\rm sys} Ω}{λ^2 \sqrt{Δν\,τ\,N_{\rm int}}} {\rm sinc}\left(\frac{m\tau}{\text{sidereal day}}\right)\]

Fields:

Tsys specifies the system temperature

τ specfies the length of a single integration

Nint specifies the total number of integrations used in the dataset

Ω is the solid angle subtended by the primary beam

Usage:

julia> model = BPJSpec.NoiseModel(1000u"K", 13u"s", 6628, 2.41u"sr")
NoiseModel(Tsys = 1000.0 K, τ = 13.0 s, Nint = 6628, Ω = 2.410 sr)

julia> model(100, 74u"MHz", 24u"kHz")
4.456470453155544 Jy

source