Power Spectrum Estimation

Power Spectrum Estimation

API

BPJSpec.q_estimator — Function.

q_estimator(mmodes, transfermatrix, covariancematrix, basis)

Evaluate the $q$ estimator:

\[q_a = v^* C^{-1} B C_a B^* C^{-1} v\]

Arguments:

mmodes or $v$ specifies the list of measured $m$-modes

transfermatrix or $B$ specifies the interferometer's response to the sky

covariancematrix or $C$ specifies the covariance of the measured $m$-modes

basis or $C_a$ is a list of angular covariance matrices that represent the change in the covariance with respect to an increase in power of each 21-cm power spectrum bin

source

BPJSpec.fisher_information — Function.

fisher_information(transfermatrix, covariancematrix, basis; iterations=10)

Compute a Monte-Carlo approximation of the Fisher information matrix.

\[F_{ab} = {\rm tr}\left( C^{-1} C_a C^{-1} C_b \right)\]

Arguments:

transfermatrix or $B$ specifies the interferometer's response to the sky

covariancematrix or $C$ specifies the covariance of the measured $m$-modes

basis or $C_a$ is a list of angular covariance matrices that represent the change in the covariance with respect to an increase in power of each 21-cm power spectrum bin

Keyword Arguments:

iterations is the number of Monte Carlo simulations to perform

source

BPJSpec.noise_bias — Function.

noise_bias(transfermatrix, covariancematrix, basis; iterations=10)

Compute a Monte-Carlo approximation of the noise bias to the quadratic estimator.

source

BPJSpec.full_rank_compress! — Function.

full_rank_compress!(output_mmodes, output_transfermatrix, output_noisematrix,
                    input_mmodes,  input_transfermatrix,  input_noisematrix;
                    progress=false)

In the case where the interferometer has more baselines than there are spherical harmonic coefficients to measure, the transfer matrix is tall and skinny. This also indicates that we have made redundant measurements that can be averaged together with no information loss.

In this routine we use the singular value decomposition (SVD) of the transfer matrix to compress the measurements. However, the SVD is just as large as the transfer matrix itself, and will take a lot of disk space to store. Therefore we will compute the SVD, compress everything with it all at once so that there is no need to store the SVD as well.

Arguments:

output_mmodes the output compressed $m$-modes

output_transfermatrix the output compressed transfer matrix

output_noisematrix the output compressed noise covariance matrix

input_mmodes the input $m$-modes that will be compressed

input_transfermatrix the input transfer matrix that will be used to generate the compression

input_noisematrix the input noise covariance matrix

Keyword Arguments:

progress if set to true, a progress bar will be displayed

source