Home

BPJSpec Documentation

BPJSpec is a 21-cm power spectrum code developed for the OVRO-LWA based on the $m$-mode analysis formalism.

BPJSpec.TransferMatrix — Type.

struct TransferMatrix <: AbstractBlockMatrix{Matrix{Complex128}, 2}

This type represents the transfer matrix of an interferometer. This matrix effectively describes how an interferometer responds to the sky, including the antenna primary beam, bandpass, and baseline distribution.

This matrix is hierarchical in the sense that we save on some computational and storage requirements by separating long baselines from short baselines.

Fields

storage contains instructions on how to read the matrix from disk

cache is used if we want to keep the matrix in memory

metadata stores the interferometer's metadata

lmax is the largest value of the $l$ quantum number used by the matrix

mmax is the largest value of the $m$ quantum number used by the matrix

source

BPJSpec.MModes — Type.

struct MModes{S} <: AbstractBlockMatrix{Vector{Complex128}, 2}

This type represents the $m$-modes measured by the interferometer.

Fields

storage contains instructions on how to read the m-modes from disk

cache is used if we want to keep the $m$-modes in memory

mmax is the largest value of the $m$ quantum number

frequencies is the list of frequencies

bandwidth is the bandwidth associated with each frequency channel

source

Internals

BPJSpec.AbstractBlockMatrix — Type.

abstract type AbstractBlockMatrix{B, N}

This type represents a (potentially enormous) block-diagonal matrix. This type is designed to be general enough to handle large matrices that fit in memory as well as enormous matrices that do not fit in memory. In principle this type can also be used to store small matrices, but it would be relatively inefficient compared to the standard Array{T, N}.

Type Parameters

B specifies the type of the blocks that compose the matrix

N specifies the number of indices used to index into the matrix blocks

Required Fields

storage contains instructions on how to read matrix blocks

cache is used if we want to read the matrix from disk and then keep it in memory for faster access.

source