Power Spectrum Estimation

Power Spectrum Estimation

API

BPJSpec.full_rank_compress! — Function.

full_rank_compress!(output_mmodes, output_transfermatrix, output_noisematrix,
                    input_mmodes,  input_transfermatrix,  input_noisematrix;
                    progress=false)

In the case where the interferometer has more baselines than there are spherical harmonic coefficients to measure, the transfer matrix is tall and skinny. This also indicates that we have made redundant measurements that can be averaged together with no information loss.

In this routine we use the singular value decomposition (SVD) of the transfer matrix to compress the measurements. However, the SVD is just as large as the transfer matrix itself, and will take a lot of disk space to store. Therefore we will compute the SVD, compress everything with it all at once so that there is no need to store the SVD as well.

Arguments:

output_mmodes the output compressed $m$-modes

output_transfermatrix the output compressed transfer matrix

output_noisematrix the output compressed noise covariance matrix

input_mmodes the input $m$-modes that will be compressed

input_transfermatrix the input transfer matrix that will be used to generate the compression

input_noisematrix the input noise covariance matrix

Keyword Arguments:

progress if set to true, a progress bar will be displayed

source